Ensure styles tree has been processed before checking its outputPath
Prior to this change, the `ModuleSourceFunnel` would assume that the
passed in `stylesTree` has been processed when its `build` hook was
called. This _happened_ to work in some cases, but since Broccoli
builder uses a DAG to decide what order to build trees in it absolutely
could find a different sort order that causes `ModuleSourceFunnel` to
run before that styles tree.

As part of this, I updated broccoli-funnel to allow `super`ing with
either a single node (prior behavior) or an array. broccoli-funnel
itself will never look at ainputs other than `inputPath[0]` but this is
useful to ensure the broccoli node graph indicates the required
ordering.