expanding embroider test coverage
I was investigating ember-css-modules behavior in addons so I expanded the embroider-app tests to handle the case of an addon that uses ember-css-modules.

In addition to the documented limitation about keeping staticComponents off, addons that use ember-css-modules also need staticAddonTrees: false for the same reason.

This expanded test coverage will be helpful for anyone who is working on changes that would allow enabling these optimizations in the future.